Added Bits, NameTag, Generations, FamilyTree
FamilyTree is the Parent level object, in charge of directing logic.

NameTag is in charge of the Square Containers that contain data (username, profile image)

Bits are the dots on the bottom of NameTags(Info Containers) - they just look nice, but their main purpose is to display connections between family members. Their job is to help organize position data and display connections between NameTags. Its nice having dots.

The generation class is in charge of displaying Visual Heigharchy. Gen0 will have a list of the generation 0, and it will be at the top, and Gen1 will be offset from the top by however many Generations there are and display its NameTag objects. --- Generation will be in charge of drawing connections between Bits (the dots). From Gen(0) -> Connection -> Gen(1) -> Connection -> Gen(2) -> Conn.. etc;